FIX: float to double

+++ lib/models/transaction.dart
@@ -1,8 +1,6 @@
-import 'dart:ffi';
-
 class Transaction {
   final String spenderUID;
-  final Float amount;
+  final double amount;
 
   Transaction(this.spenderUID, this.amount);
 }
